MCP_Servers/Nominations_SQL_Agent/requirements.txt
openai>=1.40.0
mcp>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
